+++ frontend/actions/business_actions.js
@@ -2,8 +2,7 @@ import * as APIUtil from '../util/business_api_util';
 
 export const RECEIVE_BUSINESSES = 'RECEIVE_BUSINESSES';
 export const RECEIVE_BUSINESS = 'RECEIVE_BUSINESS';
-export const RECEIVE_REVIEW = 'RECEIVE_REVIEW';
-export const RECEIVE_REVIEWS = 'RECEIVE_REVIEWS';
+
 
 export const receiveBusinesses = businesses => ({
   type: RECEIVE_BUSINESSES,
@@ -17,28 +16,7 @@ export const receiveBusiness = ({ business, reviews }) => ({
 
 });
 
-export const receiveReview = ({ body, rating, userId }) => ({
-  type: RECEIVE_REVIEW,
-  body,
-  rating,
-  userId,
-});
-export const receiveReviews = reviews  => ({
-  type: RECEIVE_REVIEWS,
-  reviews
-});
 
-export const createReview = review => dispatch => (
-  APIUtil.createReview(review).then(review => (
-    dispatch(receiveReview(review))
-  ))
-);
-
-export const fetchReviews = id => dispatch => (
-  APIUtil.fetchReviews(id).then(review => (
-    dispatch(receiveReviews(review))
-  ))
-)
 export const fetchBusinesses = data => dispatch => (
   APIUtil.fetchBusinesses(data).then(businesses => (
     dispatch(receiveBusinesses(businesses))
