+++ frontend/reducers/reviews_reducer.js
@@ -1,14 +1,15 @@
 import {
-    RECEIVE_BUSINESS,
     RECEIVE_REVIEW,
     RECEIVE_REVIEWS,
-  } from '../actions/business_actions';
-
+  } from '../actions/review_actions';
+import {
+    RECEIVE_BUSINESS
+} from '../actions/business_actions'
 const reviewsReducer = (state = {}, action) => {
     Object.freeze(state);
     switch(action.type) {
         case RECEIVE_BUSINESS:
-            return Object.assign({}, state, action.reviews);
+            return action.reviews || {};
         case RECEIVE_REVIEW:
             const { review } = action;
             return Object.assign({}, state, { [review.id]: review });
